Rename RbNaCl::Point to RbNaCl::GroupElement
